requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyodbc>=4.0.39
reportlab>=4.0.4
sqlalchemy>=2.0.0
twilio>=9.0.0
python-dotenv>=1.0.0